refactor: review usage of @emotion/css in components
- use useCss/useClasses cached version of css/cx in components
  - move stories to stories/ & remove unneeded styles from tests
- bump clsx to align with most packages (MUI)
