+++ app/package.json
@@ -20,7 +20,7 @@
     "@mui/material": "^5.13.5",
     "@types/react-grid-layout": "^1.3.4",
     "arquero": "^5.2.0",
-    "clsx": "^1.2.1",
+    "clsx": "^2.0.0",
     "formik": "^2.4.2",
     "i18next": "^22.5.1",
     "i18next-browser-languagedetector": "^7.0.2",
