+++ packages/core/package.json
@@ -54,7 +54,7 @@
     "@react-stately/datepicker": "^3.4.1",
     "@types/react-table": "^7.7.14",
     "attr-accept": "^2.2.2",
-    "clsx": "^1.2.1",
+    "clsx": "^2.0.0",
     "dayjs": "^1.11.8",
     "detect-browser": "^5.3.0",
     "embla-carousel-react": "^8.0.0-rc07",
