+++ packages/core/src/components/Drawer/Drawer.test.tsx
@@ -1,7 +1,6 @@
 import { useState } from "react";
 import userEvent from "@testing-library/user-event";
 import { render } from "@testing-library/react";
-import { css } from "@emotion/css";
 
 import {
   HvButton,
@@ -14,25 +13,6 @@ import { HvDrawer } from "./Drawer";
 const Sample = ({ startingOpen }: { startingOpen: boolean }) => {
   const [open, setOpen] = useState(startingOpen);
 
-  const drawerWidth = "60%";
-
-  const classes = {
-    drawerPaper: css({
-      width: `calc(100% - ${drawerWidth})`,
-      overflow: "unset", // we want scroll content, not scroll paper
-    }),
-    drawerContent: css({
-      // maximize space for content
-      flex: 1,
-      overflow: "auto",
-    }),
-    drawerTitle: css({
-      flexGrow: 0,
-      flexShrink: 0,
-      flexBasis: "auto",
-    }),
-  };
-
   return (
     <div>
       <HvButton
@@ -42,18 +22,9 @@ const Sample = ({ startingOpen }: { startingOpen: boolean }) => {
       >
         Open dialog
       </HvButton>
-      <HvDrawer
-        id="drawer-test"
-        open={open}
-        classes={{
-          paper: classes.drawerPaper,
-        }}
-        onClose={() => setOpen(false)}
-      >
-        <HvDialogTitle classes={{ root: classes.drawerTitle }}>
-          Lorem Ipsum
-        </HvDialogTitle>
-        <HvDialogContent className={classes.drawerContent}>
+      <HvDrawer id="drawer-test" open={open} onClose={() => setOpen(false)}>
+        <HvDialogTitle>Lorem Ipsum</HvDialogTitle>
+        <HvDialogContent>
           {`Cras mattis consectetur purus sit amet fermentum. 
                   Cras justo odio, dapibus ac facilisis in, egestas eget quam. 
                   Morbi leo risus, porta ac consectetur ac, vestibulum at eros.
